Rename Character modules all lowercase for review #677.
