+++ default/python/AI/ProductionAI.py
@@ -4,7 +4,7 @@ import random
 import freeOrionAIInterface as fo  # pylint: disable=import-error
 import AIDependencies
 import AIstate
-from character.Character import Aggression
+from character.character_module import Aggression
 import FleetUtilsAI
 import FreeOrionAI as foAI
 import PlanetUtilsAI
